+++ src/screens/About/index.tsx
@@ -1,16 +1,6 @@
 import React, { FC } from 'react';
-import { Text, ScrollView, Button, Alert, Dimensions, Image, } from 'react-native';
-import styled from 'styled-components/native';
+import { Text, ScrollView, Image, } from 'react-native';
 import {globalStyles} from '../../styles/globalStyle';
-import Colors from '../../constans/Colors';
-import { red } from 'color-name';
-
-const ListText = styled.Text`
-    margin: 120px 20px;
-    font-size: 16px;
-    color: ${Colors.black};
-`;
-
 
 interface IAboutProps {
     myProps: string;
@@ -20,8 +10,8 @@ const Lista="Cos";
 const About: FC<IAboutProps> = props => {
     return (
         <ScrollView style={globalStyles.container}>
-            <Image style={{width:100, height:100, marginLeft:10,marginTop:40 ,borderRadius:50}} source={require('../../assets/KR-VEC-LOGO.png')}/>
-            <Text  style={{margin:10,fontSize:20}}>
+            <Image style={globalStyles.roundImage} source={require('../../assets/KR-VEC-LOGO.png')}/>
+            <Text  style={{margin:10,paddingBottom:40,fontSize:20}}>
                 Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum
                 Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum
                 Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum