+++ CHANGELOG.md
@@ -4,12 +4,79 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - 2020-02-25
+## [1.0.0] - 2020-03-10
 
 ### Changes
-
+    
     - 
 
 ### Added
 
-    - Initial app config
+    - Welcome screen
+    - List screen
+    - About screen
+
+## [1.0.1] - 2020-03-15
+
+### Changes
+    
+    - 
+
+### Added
+
+    - Navigation Bar
+
+## [1.0.2] - 2020-03-16
+
+### Changes
+    
+    - Welcome screen
+
+### Added
+
+    - Layout in Welcome screen
+
+## [1.0.3] - 2020-03-20
+
+### Changes
+    
+    - Welcome screen
+    - List screen
+    - About screen
+
+### Added
+
+    - Layout in Welcome screen
+    - Layout in List screen
+    - Layout in About screen
+
+## [1.0.4] - 2020-04-1
+
+### Changes
+    
+    - Splash screen
+
+### Added
+
+    - AddNote screen
+
+## [1.0.5] - 2020-04-2
+
+### Changes
+    
+    - Modal
+
+### Added
+
+    - Formik
+    - List in List screen
+
+## [1.0.5] - 2020-04-2
+
+### Changes
+    -Notes are added to List
+
+### Added
+
+    - Formik
+    - List in List screen